Some Improvements to the Gitian Verification Script
Keys are no longer imported automatically, use `--import-keys`
as an argument to import all public keys in the `gitian-pubkeys`
directory.

Keys are no longer refreshed automatically, use `--refresh-keys`
as an argument to refresh all public keys in the gpg keyring from
the sks server.

The verification step now uses the current working directory instead
of a hardcoded name in case the user renamed the gitian.sigs.

The verification now checks for the existance of each gitian.sigs
directory before running.

The user can specify custom monero and gitian-builder directory
locations. By default monero and gitian-builder are assumed on the same
level as the gitian.sigs repository.